requirements.txt
fastapi
uvicorn
pynacl
pydantic
pydantic-settings
